refactor: Layout 생성 코드 분리
